Adding Error Codes 205, 206, 207, 208 and 209
Adding error Code  205, 206, 207, 208 and 209 for the Years in CSV Datei. Check if the Year Value is = 0, print error code 205, if the value is under 4 numbers long print error code 206, error code 207 if value is smaller than the year value before, error code 208 if the value in Year is bigger than the next year value, error code 209 if the value is the same as the smaller or the bigger value. Block these actions to provide broken data